+++ crates/koto/examples/rust_function.rs
@@ -1,4 +1,4 @@
-use koto::{prelude::*, runtime};
+use koto::prelude::*;
 
 fn main() {
     let script = "
@@ -21,7 +21,7 @@ print plus 10, 20
     koto.compile_and_run(script).unwrap();
 }
 
-fn say_hello(ctx: &mut CallContext) -> runtime::Result<KValue> {
+fn say_hello(ctx: &mut CallContext) -> koto::Result<KValue> {
     match ctx.args() {
         [] => println!("Hello?"),
         [KValue::Str(name)] => println!("Hello, {name}"),